+++ apps/web/app/(full-layout)/roadmaps/generate/_components/step-quiz.tsx
@@ -12,7 +12,7 @@ import type {
 } from '@/app/(full-layout)/roadmaps/generate/_types/onboarding';
 
 import { LoadingState } from '@/components/shared/loading-state';
-import { onboardingService } from '@/services/onboarding-service';
+import { onboardingService } from '@/services/onboarding.service';
 
 import { PersonalizedQuestionsPanel } from './personalized-questions-panel';
