+++ apps/web/components/providers/auth-provider.tsx
@@ -5,7 +5,7 @@ import { createContext, useCallback, useContext, useMemo, useState } from 'react
 import type { AuthUser } from '@/types/auth';
 import type { SignInValues, SignUpValues } from '@/validations/auth.schema';
 
-import { authService } from '@/services/auth-service';
+import { authService } from '@/services/auth.service';
 import { normalizeUser } from '@/utils/user';
 
 export interface AuthContextValue {
